refactor: restructure package into five domain subpackages
- Replace extraction/ with pathways/ combining O*NET schemas, credentials, clustering types, graph edges, and career pathway graph
- Move PDF reader into matching/ as the resume input stage
- Consolidate report/ into display/ with figures.py, tables.py, and layout.py
- Add Marimo layout helpers for stat rows, filtered accordions, and conditional tables
- Extract gap and demonstrated row builders from notebook into display/tables.py
- Slim pipeline/schemas.py to PipelineConfig and Encoder only
- Move Corpus onto collection/schemas.py alongside Posting
- Decouple CareerPathwayGraph from PipelineConfig via individual scalar parameters
- Delete PatternBundle (zero references)
- Replace NaN self-inequality check with pd.isna in collector
- Remove defensive empty-frames guard from scraper concat
- Fix Neighborhood import bug in figures.py
- Restructure tests to mirror new source layout
