+++ src/chalkline/collection/schemas.py
@@ -1,13 +1,15 @@
 """
 Schemas for corpus collection.
 
-Defines the `Posting` schema with deterministic composite key deduplication.
+Defines the `Posting` schema with deterministic composite key deduplication
+and the `Corpus` container that indexes postings for deterministic encoding.
 """
 
-from datetime import date
-from pydantic import BaseModel, BeforeValidator, Field, model_validator
-from slugify  import slugify
-from typing   import Annotated, Self
+from dataclasses import dataclass, field
+from datetime    import date
+from pydantic    import BaseModel, Field, model_validator
+from slugify     import slugify
+from typing      import Annotated, Self
 
 
 class Posting(BaseModel, extra="forbid"):
@@ -20,16 +22,13 @@ class Posting(BaseModel, extra="forbid"):
     """
 
     company     : str
-    date_posted : Annotated[
-                      date | None,
-                      BeforeValidator(lambda v: v[:10] if isinstance(v, str) else v)
-                  ]
+    date_posted : date | None
     description : Annotated[str, Field(min_length=50)]
     source_url  : str
     title       : str
 
     date_collected : date       = Field(default_factory=date.today)
-    id             : str       = ""
+    id             : str        = ""
     location       : str | None = None
 
     @model_validator(mode="after")
@@ -45,3 +44,29 @@ class Posting(BaseModel, extra="forbid"):
                 f"{self.date_posted.isoformat() if self.date_posted else 'undated'}"
             )
         return self
+
+
+@dataclass
+class Corpus:
+    """
+    Filtered posting corpus with eagerly-derived key ordering.
+
+    Keeps `Posting` objects intact so downstream consumers access
+    `.description` and `.title` through the posting rather than through
+    parallel dicts. The sorted key list is computed once at construction for
+    deterministic encoding and manifest ordering.
+    """
+
+    postings    : dict[str, Posting]
+    posting_ids : list[str] = field(init=False)
+
+    def __post_init__(self):
+        self.posting_ids = sorted(self.postings)
+
+    @property
+    def descriptions(self) -> list[str]:
+        """
+        Posting descriptions in deterministic sorted-key order for sentence
+        encoding.
+        """
+        return [self.postings[pid].description for pid in self.posting_ids]
